// packages/aspida/__tests__/apis/v1.1/2/_hogeId@HogeId/test-4/fuga aa.ts
export interface Methods {
  get: {
    query: { aa?: number }
    resData: { id: number }
  }

  post: {
    query: { aa: number }
    reqData?: { name: string }
    resData: { id: number }
  }

  put: {
    query: { aa: number }
    resData: { id: number }
  }

  delete: {
    query: { aa: number }
    reqData: { name: string }
    resData: { id: number }
  }
}